+++ primefaces-showcase/src/main/java/org/primefaces/showcase/view/button/BasicView.java
@@ -24,6 +24,7 @@
 package org.primefaces.showcase.view.button;
 
 import java.io.Serializable;
+import java.util.concurrent.TimeUnit;
 import javax.faces.view.ViewScoped;
 import javax.inject.Named;
 
@@ -31,8 +32,8 @@ import javax.inject.Named;
 @ViewScoped
 public class BasicView implements Serializable {
 
-    public void sleep(int msec) throws InterruptedException {
-        Thread.sleep(msec);
+    public void sleep() throws InterruptedException {
+        TimeUnit.SECONDS.sleep(1);
     }
 
 }