+++ docs/12_0_0/gettingstarted/whatsnew.md
@@ -5,7 +5,9 @@ This page contains a list of big features. Please check the GitHub issues for al
 ## 12.0.0
 
 * AutoComplete: added `lazyModel` attribute for fetching autocomplete-suggestions (instead of completeMethod).
-* CommandButton: added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
+* CommandButton:
+  * added `disableOnAjax` attribute to disable the button during Ajax requests triggered by it.
+  * added `inlineAjaxStatus` attribute to show a loading icon in the button during Ajax requests triggered by it.
 * CommandLink
   * Now has a widget and can be enabled and disabled by its widget.
   * Added `disableOnAjax` attribute to disable the link during Ajax requests triggered by it.